fix: create assessment group on install
