+++ education/patches.txt
@@ -2,6 +2,6 @@
 education.patches.v14_0.create_student_party_type
 
 [post_model_sync]
-education.patches.v14_0.create_parent_assessment_group #01-12-2022
+education.patches.v14_0.create_parent_assessment_group #01-01-2022
 education.patches.v14_0.lms_deprecation_message #22-12-2022
 education.patches.v14_0.student_name
