+++ education/patches/v14_0/create_parent_assessment_group.py
@@ -1,12 +1,8 @@
 import frappe
+from education.install import create_parent_assessment_group
 
 def execute():
-    if not frappe.db.exists("Assessment Group", "All Assessment Groups"):
-        frappe.get_doc({
-            "doctype": "Assessment Group",
-            "assessment_group_name": "All Assessment Groups",
-            "is_group": 1
-        }).insert(ignore_mandatory=True)
+    create_parent_assessment_group()
 
     if frappe.db.exists("Assessment Group", "undefined"):
-        frappe.delete_doc("Assessment Group", "undefined")
+        frappe.delete_doc("Assessment Group", "undefined")
